requirements.txt
borsapy==0.8.3
httpx>=0.28.1
openai>=1.73.0
python-dotenv>=1.0.1
